third_party/blink/web_tests/wpt_internal/task-tracking/track-postmessage.html

<!DOCTYPE HTML>
<html>
<head>
<meta charset="utf-8">
<title>Verify that postmessage tasks can be properly tracked.</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="resources/task-ids.js"></script>
</head>
<body>
<script>
const generatePromise = (scriptId, receiver) => {
  return new Promise((resolve, reject) => {
    const handler = () => {
      try {
        assert_equals(scheduler.taskId, scriptId);
        resolve();
      } catch(e) {
        reject(e);
      }
    }
    receiver.onmessage = handler;
  });
}

promise_test(t => {
  const scriptId = initializeTaskId();
  const channel1 = new MessageChannel();
  const channel2 = new MessageChannel();

  const promise1 = generatePromise(scriptId, channel1.port1);
  const promise2 = generatePromise(scriptId, channel2.port1);
  channel1.port2.postMessage("hello");
  channel2.port2.postMessage("hello");
  return Promise.all([promise1, promise2]);
}, "Same document MessagePort postMessage task properly detects its parent.");

promise_test(t => {
  const scriptId = initializeTaskId();
  const p = generatePromise(scriptId, window);
  window.postMessage("hello", "*");
  return p;
}, "Same document window.postMessage task properly detects its parent.");

</script>
</body>
</html>
